Complete the message list screen
With Groupie another list is set, the data are simple dummies data provide in the data samples package. Custom adapters are created in order to manage the binding between the view and the data. The toolbar title is styled too. Glide dependencies is used to display round image from the internet or png bitmap.